feature(nav): nav v2 - in progress
