+++ client/src/components/Navbar/Navbar.js
@@ -1,100 +1,6 @@
-import React, { useState } from "react";
-import styled from "styled-components";
-import { Link } from "react-router-dom";
-
-import useWindowSize from "../../utils/HelperFunc/WindowSize.js";
-
-const navBarStyling = {
-  navBarTransparent: {
-    zIndex: "999",
-    width: "100%",
-    height: "80px",
-    position: "absolute",
-    borderRadius: "0 0 35px 35px",
-    display: "flex",
-    justifyContent: "space-between",
-    alignItems: "center",
-    padding: "20px 0",
-    backgroundColor: "transparent",
-  },
-  navBarWhite: {
-    zIndex: "999",
-    width: "100%",
-    height: "50px",
-    position: "fixed",
-    borderRadius: "0 0 35px 35px",
-    boxShadow: "0 10px 35px rgb(0 30 47 / 10%)",
-    display: "flex",
-    justifyContent: "space-between",
-    alignItems: "center",
-    padding: "20px 0",
-    backgroundColor: "white",
-  },
-};
-
-const StyledLogoLink = styled(Link)`
-  font-size: 24px;
-  font-weight: 700;
-  color: ${(props) => (props.navColor ? "black" : "white")};
-  margin: 0 10px;
-  padding: 5px 0;
-  text-decoration: none;
-  &:hover {
-    border-bottom: ${(props) =>
-      props.navColor ? "2px solid rgb(25, 200, 140)" : "2px solid white"};
-  }
-  @media (max-width: 1200px) {
-    display: none;
-  }
-`;
-
-const StyledLink = styled(Link)`
-  font-size: 18px;
-  font-weight: 500;
-  color: ${(props) => (props.navColor ? "black" : "white")};
-  margin: 0 10px;
-  padding: 5px 0;
-  text-decoration: none;
-  &:hover {
-    border-bottom: ${(props) =>
-      props.navColor ? "2px solid rgb(25, 200, 140)" : "2px solid white"};
-  }
-  @media (max-width: 768px) {
-    flex-direction: column;
-  }
-`;
-
-const StyleButtonLink = styled(Link)`
-  text-decoration: none;
-  color: ${(props) => (props.navColor ? "white" : "rgb(25, 200, 140)")};
-  font-size: 16px;
-  @media (max-width: 768px) {
-    flex-direction: column;
-  }
-`;
-
-const Button = styled.div`
-  width: 120px;
-  height: 35px;
-  padding: 10px 12px;
-  border: none;
-  border-radius: 50px;
-  display: flex;
-  font-size: 16px;
-  font-weight: 700;
-  align-items: center;
-  justify-content: center;
-  background-color: ${(props) =>
-    props.navColor ? "rgb(25, 200, 140)" : "white"};
-  &:hover {
-    color: rgb(28, 166, 118);
-    background-color: ${(props) =>
-      props.navColor ? "rgb(28,166,118)" : "white"};
-    transition-delay: 100ms;
-    transition: transform 100ms ease-in-out 25ms;
-    transform: translateY(3px);
-  }
-`;
+import React, { useState, useEffect } from "react";
+import { Primary, Secondary } from "../Navbar/index.js";
+import { primaryNavStyling, secondaryNavStyling } from "./styles.js";
 
 const NavBar = () => {
   // State to handle conditional render && event handling
@@ -117,70 +23,20 @@ const NavBar = () => {
       {windowSize >= 1200 ? (
         <div
           style={
-            navColor
-              ? navBarStyling.navBarWhite
-              : navBarStyling.navBarTransparent
+            !navColor ? primaryNavStyling.transparent : primaryNavStyling.white
           }
         >
-          {/* Logo */}
-          <div style={{ width: "226px", paddingLeft: "80px" }}>
-            <StyledLogoLink navColor={navColor} to="/">
-              Logo
-            </StyledLogoLink>
-          </div>
-          {/* Nav Items */}
-          <div style={{ display: "flex" }}>
-            <div style={{ justifyContent: "center" }}>
-              <StyledLink navColor={navColor} to="/">
-                NAV
-              </StyledLink>
-              <StyledLink navColor={navColor} to="/">
-                NAV
-              </StyledLink>
-              <StyledLink navColor={navColor} to="/">
-                NAV
-              </StyledLink>
-              <StyledLink navColor={navColor} to="/">
-                NAV
-              </StyledLink>
-              <StyledLink navColor={navColor} to="/">
-                NAV
-              </StyledLink>
-            </div>
-          </div>
-          <div
-            style={{
-              display: "flex",
-              alignItems: "center",
-              paddingRight: "80px",
-            }}
-          >
-            {/* Login and Sign up button */}
-            <div>
-              <StyledLink
-                style={{ marginRight: "20px" }}
-                navColor={navColor}
-                to="/login"
-              >
-                Login In
-              </StyledLink>
-            </div>
-            <Button navColor={navColor}>
-              <StyleButtonLink navColor={navColor} to="/">
-                Join For Free
-              </StyleButtonLink>
-            </Button>
-          </div>
+          <Primary navColor={navColor} />
         </div>
       ) : (
         <div
           style={
-            navColor
-              ? navBarStyling.navBarWhite
-              : navBarStyling.navBarTransparent
+            !navColor
+              ? secondaryNavStyling.transparent
+              : secondaryNavStyling.white
           }
         >
-          hello
+          <Secondary navColor={navColor} />
         </div>
       )}
     </>
@@ -188,3 +44,30 @@ const NavBar = () => {
 };
 
 export default NavBar;
+
+//--------------------------- Window Size Helper Function -------------------------------------
+
+const useWindowSize = () => {
+  const isWindowClient = typeof window === "object";
+
+  const [windowSize, setWindowSize] = useState(
+    isWindowClient ? window.innerWidth : undefined
+  );
+
+  useEffect(() => {
+    //a handler which will be called on change of the screen resize
+    const setSize = () => {
+      setWindowSize(window.innerWidth);
+    };
+
+    if (isWindowClient) {
+      //register the window resize listener
+      window.addEventListener("resize", setSize);
+
+      //un-register the listener
+      return () => window.removeEventListener("resize", setSize);
+    }
+  }, [isWindowClient, setWindowSize]);
+
+  return windowSize;
+};
