+++ client/src/components/Navbar/index.js
@@ -1,0 +1,4 @@
+import Primary from "./Primary/Primary.js";
+import Secondary from "./Secondary/Secondary.js";
+
+export { Primary, Secondary };
