+++ client/src/components/Navbar/styles.js
@@ -1,0 +1,55 @@
+export const primaryNavStyling = {
+  transparent: {
+    zIndex: "999",
+    width: "100%",
+    height: "80px",
+    position: "absolute",
+    borderRadius: "0 0 35px 35px",
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+    padding: "20px 0",
+    backgroundColor: "transparent",
+  },
+  white: {
+    zIndex: "999",
+    width: "100%",
+    height: "50px",
+    position: "fixed",
+    borderRadius: "0 0 35px 35px",
+    boxShadow: "0 10px 35px rgb(0 30 47 / 10%)",
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+    padding: "20px 0",
+    backgroundColor: "white",
+  },
+};
+
+export const secondaryNavStyling = {
+  transparent: {
+    zIndex: "999",
+    width: "100%",
+    height: "80px",
+    position: "absolute",
+    borderRadius: "0 0 35px 35px",
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+    padding: "20px 0",
+    backgroundColor: "transparent",
+  },
+  white: {
+    zIndex: "999",
+    width: "100%",
+    height: "50px",
+    position: "fixed",
+    borderRadius: "0 0 35px 35px",
+    boxShadow: "0 10px 35px rgb(0 30 47 / 10%)",
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+    padding: "20px 0",
+    backgroundColor: "white",
+  },
+};
